+++ packages/webgpu-waveform-react/src/useWaveformRenderer.tsx
@@ -1,6 +1,5 @@
-import { RefObject, useEffect, useState } from "react";
+import { useEffect, useState } from "react";
 import { GPUWaveformRenderer } from "webgpu-waveform";
-import { nullthrows } from "./useWebGPU";
 
 type RendererStatus =
   | { status: "initializing" }
@@ -10,27 +9,19 @@ type RendererStatus =
       instance: GPUWaveformRenderer;
     };
 
-export function useWaveformRenderer(
-  canvasRef: RefObject<HTMLCanvasElement>,
-  audioBuffer: AudioBuffer
-): RendererStatus {
+export function useWaveformRenderer(audioBuffer: AudioBuffer): RendererStatus {
   const [renderer, setRenderer] = useState<RendererStatus>({
     status: "initializing",
   });
 
   useEffect(() => {
-    const canvas = nullthrows(
-      canvasRef.current,
-      "expected canvas to not be nil"
-    );
     const channelData = audioBuffer.getChannelData(0);
-
-    GPUWaveformRenderer.create(canvas, channelData)
+    GPUWaveformRenderer.create(channelData)
       .then((waveformRenderer: GPUWaveformRenderer) =>
         setRenderer({ status: "ready", instance: waveformRenderer })
       )
       .catch((err: any) => setRenderer({ status: "error", error: err }));
-  }, [audioBuffer, canvasRef]);
+  }, [audioBuffer]);
 
   return renderer;
 }